Add APPEND command
If key already exists and is a string, this command appends the value at the end
of the string. If key does not exist it is created and set as an empty string, then its appended with given value. So, APPEND can be used as SET in its special cases.