Add INCRBYFLOAT command
Increment the string representing a floating point number stored at key by the specified increment. If the key does not exist, it is set to 0 before performing the operation. An error is returned if one of the following conditions occur:
The key contains a value of the wrong type (not a string).
The current key content or the specified increment are not parsable as a double precision floating point number.